requirements.txt
pandas
pyyaml==6.0.1
pytest
python-dotenv
ftfy
openai
